requirements.txt
# 核心机器学习库
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# 深度学习框架 (针对NVIDIA 4090 + CUDA 12.4优化)
# PyTorch - 支持CUDA 12.4的最新稳定版本
torch>=2.1.0,<2.3.0
torchvision>=0.16.0,<0.18.0
torchaudio>=2.1.0,<2.3.0
# 如需CUDA支持，请安装: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124

# TensorFlow - 支持CUDA 12.4的版本
tensorflow>=2.15.0,<2.16.0
# TensorFlow GPU支持说明: TensorFlow 2.15+原生支持CUDA 12.4
keras>=3.0.0

# 数据可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.3.0

# 数据处理
openpyxl>=3.0.0
xlrd>=2.0.0
h5py>=3.1.0
pillow>=8.0.0

# 模型解释性
shap>=0.39.0
lime>=0.2.0
eli5>=0.11.0

# 模型优化
optuna>=2.8.0
hyperopt>=0.2.5
ray[tune]>=1.4.0

# 模型部署
flask>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=0.84.0

# 实验跟踪
mlflow>=1.18.0
wandb>=0.12.0
tensorboard>=2.6.0

# 数据验证
great-expectations>=0.13.0
pydantic>=1.8.0

# 工具库
tqdm>=4.61.0
click>=8.0.0
python-dotenv>=0.19.0
pyyaml>=5.4.0
joblib>=1.0.0
dill>=0.3.0

# 开发工具
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.6.0
flake8>=3.9.0
mypy>=0.910

# Jupyter支持
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# 其他常用库
requests>=2.25.0
beautifulsoup4>=4.9.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pymongo>=3.12.0
redis>=3.5.0